(feat) Implement knowledge repo bar to render user's favorite guides
